Moved Platform version determination to static
