+++ src/NzbDrone.Mono.Test/EnvironmentInfo/MonoPlatformInfoFixture.cs
@@ -1,14 +1,14 @@
 using System;
 using FluentAssertions;
 using NUnit.Framework;
-using NzbDrone.Mono.EnvironmentInfo;
+using NzbDrone.Common.EnvironmentInfo;
 using NzbDrone.Test.Common;
 
 namespace NzbDrone.Mono.Test.EnvironmentInfo
 {
     [TestFixture]
     [Platform("Mono")]
-    public class MonoPlatformInfoFixture : TestBase<MonoPlatformInfo>
+    public class MonoPlatformInfoFixture : TestBase<PlatformInfo>
     {
         [Test]
         public void should_get_framework_version()
